# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Live Restaurant Queue & Pre-Order Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Restaurant } from './restaurant';

export interface AuthResponse {
  restaurant: Restaurant;
  token: string;
}