# lib/api-zod/src/generated/types/menuItemUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Live Restaurant Queue & Pre-Order Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItemUpdate {
  name?: string;
  price?: number;
  category?: string;
  isAvailable?: boolean;
}