# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Live Restaurant Queue & Pre-Order Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  queueEntryId: number;
  restaurantId: number;
  customerName?: string;
  queueNumber?: number;
  items: OrderItem[];
  status: OrderStatus;
  totalAmount: number;
  paymentStatus: OrderPaymentStatus;
  createdAt: string;
}