# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Live Restaurant Queue & Pre-Order Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from './orderInputItemsItem';

export interface OrderInput {
  sessionToken: string;
  items: OrderInputItemsItem[];
}