# lib/api-zod/src/generated/types/orderStatusUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Live Restaurant Queue & Pre-Order Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatusUpdatePaymentStatus } from './orderStatusUpdatePaymentStatus';
import type { OrderStatusUpdateStatus } from './orderStatusUpdateStatus';

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  paymentStatus?: OrderStatusUpdatePaymentStatus;
}