# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Live Restaurant Queue & Pre-Order Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Restaurant {
  id: number;
  name: string;
  ownerName: string;
  email: string;
  phone: string;
  address: string;
  cuisineType: string;
  /** @nullable */
  photo?: string | null;
  avgWaitPerTableMins: number;
  queueOpen: boolean;
  preOrderThreshold?: number;
  createdAt: string;
}